Updated project to the latest LeapMotion Linux SDK
This seems to be the minimal set of changes needed to get this to run on
the latest LeapMotion SDK. There were a couple of
pointlessly-deprecated APIs that needed to be changed as well.

It won't run without copying the two .so files from either x64 or x86,
as required. Not sure if there's a better way to do that.

Also, this was tested only with Linux. There's probably Windows and Mac
SDK files that should be included. Maybe someone with one of those
machines wants to merge those in?

It's quite possible that I just missed the easy way to run the project
as-is on Linux. It seems more likely that Linux support didn't exist 11
months ago, though, and this was unavoidable. Linux doesn't even have
the screen api yet, so the mouse example doesn't work. I hope that this
is helpful to someone anyway!